requirements.txt
alembic==1.4.2
appdirs==1.4.4
asn1crypto==1.4.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
decorator==4.4.2
env==0.1.0
Flask==1.1.2
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
flask-paginate==0.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-SSE==0.2.1
Flask-thumbnails==1.0.6
Flask-WTF==0.14.3
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
humanize==2.6.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.19.0
mysql-connector-python==8.0.21
packaging==20.4
pbr==5.5.0
Pillow==7.2.0
pip-review==1.1.0
protobuf==3.13.0
pycparser==2.20
PyMySQL==0.10.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
qrcode==6.1
redis==3.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
sqlalchemy-filters==0.12.0
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
Tempita==0.5.2
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.4
zope.interface==5.1.0
